+++ broker-auth-local/src/main/java/org/aktin/broker/auth/cred2/http/AuthEndpoint.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2;
+package org.aktin.broker.auth.cred2.http;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -13,6 +13,8 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import org.aktin.broker.auth.cred2.auth.Token;
+import org.aktin.broker.auth.cred2.auth.TokenManager;
 import org.aktin.broker.rest.Authenticated;
 import org.aktin.broker.rest.RequireAdmin;
 import org.aktin.broker.server.auth.HttpBearerAuthentication;
