+++ broker-auth-local/src/main/java/org/aktin/broker/auth/cred2/http/Credentials.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2;
+package org.aktin.broker.auth.cred2.http;
 
 import java.io.Serializable;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -7,9 +7,10 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
-public class Status implements Serializable {
+public class Credentials implements Serializable {
 
   private static final long serialVersionUID = 1L;
 
-  public long issued;
+  public String username;
+  public String password;
 }
