+++ broker-auth-local/src/main/java/org/aktin/broker/auth/cred2/http/Status.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2;
+package org.aktin.broker.auth.cred2.http;
 
 import java.io.Serializable;
 import javax.xml.bind.annotation.XmlAccessType;
@@ -7,10 +7,9 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
-public class Credentials implements Serializable {
+public class Status implements Serializable {
 
   private static final long serialVersionUID = 1L;
 
-  public String username;
-  public String password;
+  public long issued;
 }
