+++ broker-auth-local/src/main/java/org/aktin/broker/auth/cred2/repository/User.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.cred2;
+package org.aktin.broker.auth.cred2.repository;
 
 import java.time.Instant;
 import java.util.Objects;
