+++ broker-auth-local/src/main/java/org/aktin/broker/auth/cred2/repository/UserRepository.java
@@ -1,10 +1,10 @@
-package org.aktin.broker.auth.cred2;
+package org.aktin.broker.auth.cred2.repository;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
 
-public interface UserDao {
+public interface UserRepository {
 
   User find(Connection c, String username) throws SQLException;
